# implementations/go/examples/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the DDD-based Account domain example service (Account/Card/Payment/Refund/Credential Bounded Contexts). Generated from the @-annotations above each handler in internal/interface/http/ by swag (docs/architecture/api-documentation.md) — never hand-edited.",
        "title": "Account Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated requester's payments, newest first, paginated with `page`/`take`. Out-of-range `page`/`take` values fall back to their defaults rather than failing.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "List the requester's payments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number, starting at 0",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "take",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The payment list was found.",
                        "schema": {
                            "$ref": "#/definitions/http.GetPaymentsResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Charges an active card linked to an active account with sufficient balance. The account balance is debited asynchronously once the payment completes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create a payment",
                "parameters": [
                    {
                        "description": "Card to charge and amount",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The payment was created.",
                        "schema": {
                            "$ref": "#/definitions/http.PaymentResponse"
                        }
                    },
                    "400": {
                        "description": "One of: request validation failed (`VALIDATION_FAILED`), the card is not active (`PAYMENT_REQUIRES_ACTIVE_CARD`), the linked account is not active (`PAYMENT_REQUIRES_ACTIVE_ACCOUNT`), or the account balance is insufficient (`INSUFFICIENT_BALANCE`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "One of: no card exists with the given `cardId` (`LINKED_CARD_NOT_FOUND`), or the card's linked account could not be found (`LINKED_ACCOUNT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/{paymentId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the payment only if it belongs to the authenticated requester.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Look up a payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The payment ID",
                        "name": "paymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The payment was found.",
                        "schema": {
                            "$ref": "#/definitions/http.PaymentResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No payment exists with the given `paymentId` for this requester (`PAYMENT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/{paymentId}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels a completed payment before it is refunded.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Cancel a payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The payment ID",
                        "name": "paymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cancellation reason",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CancelPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The payment was cancelled."
                    },
                    "400": {
                        "description": "One of: request validation failed (`VALIDATION_FAILED`), or only a completed payment can be cancelled (`PAYMENT_CANCEL_REQUIRES_COMPLETED_PAYMENT`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No payment exists with the given `paymentId` (`PAYMENT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/{paymentId}/refunds": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the refunds requested against a payment, newest first, paginated with `page`/`take`. Out-of-range `page`/`take` values fall back to their defaults rather than failing.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "List a payment's refunds",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The payment ID",
                        "name": "paymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number, starting at 0",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "take",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The refund list was found.",
                        "schema": {
                            "$ref": "#/definitions/http.GetRefundsResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No payment exists with the given `paymentId` (`PAYMENT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requests a refund for a payment. Eligibility is judged synchronously (`APPROVED`/`REJECTED`); an approved refund is credited back to the account asynchronously.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Request a refund",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The payment ID",
                        "name": "paymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Refund amount and reason",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.RequestRefundRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The refund request was recorded (its `status` may still be `REJECTED` — check the response body, not just the HTTP status).",
                        "schema": {
                            "$ref": "#/definitions/http.RefundResponse"
                        }
                    },
                    "400": {
                        "description": "Request validation failed (`VALIDATION_FAILED`) — e.g. a non-positive amount or missing reason.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No payment exists with the given `paymentId` (`PAYMENT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.CancelPaymentRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "description": "Why the payment is being cancelled.",
                    "type": "string"
                }
            }
        },
...
        "http.CreatePaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "The amount to charge. Must be a positive integer.",
                    "type": "integer"
                },
                "cardId": {
                    "description": "The active card to charge. Must be linked to an active account with sufficient balance.",
                    "type": "string"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "A stable, machine-readable error code the client can branch on (e.g. `ACCOUNT_NOT_FOUND`). Unlike `message`, this never changes wording.",
                    "type": "string"
                },
                "error": {
                    "description": "The standard HTTP status text for `statusCode` (e.g. `Not Found`).",
                    "type": "string"
                },
                "message": {
                    "description": "A human-readable description of the error.",
                    "type": "string"
                },
                "statusCode": {
                    "description": "The HTTP status code.",
                    "type": "integer"
                }
            }
        },
...
        "http.GetPaymentsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "The total number of payments matching the filter (not just the current page's size).",
                    "type": "integer"
                },
                "payments": {
                    "description": "The requester's payments, newest first.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.PaymentResponse"
                    }
                }
            }
        },
        "http.GetRefundsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "The total number of refunds matching the filter (not just the current page's size).",
                    "type": "integer"
                },
                "refunds": {
                    "description": "The refunds requested against the payment, newest first.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.RefundResponse"
                    }
                }
            }
        },
...
        "http.PaymentResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "The account debited (asynchronously) for this payment.",
                    "type": "string"
                },
                "amount": {
                    "description": "The charged amount.",
                    "type": "integer"
                },
                "cardId": {
                    "description": "The card used for this payment.",
                    "type": "string"
                },
                "createdAt": {
                    "description": "When the payment was created.",
                    "type": "string"
                },
                "ownerId": {
                    "description": "The payment owner's user ID.",
                    "type": "string"
                },
                "paymentId": {
                    "description": "The payment's ID.",
                    "type": "string"
                },
                "status": {
                    "description": "The payment status (`COMPLETED` or `CANCELLED`).",
                    "type": "string"
                }
            }
        },
        "http.RefundResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "The requested refund amount.",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "When the refund was requested.",
                    "type": "string"
                },
                "decisionNote": {
                    "description": "Why the refund was rejected. Only present when `status` is `REJECTED`.",
                    "type": "string"
                },
                "paymentId": {
                    "description": "The payment this refund was requested against.",
                    "type": "string"
                },
                "reason": {
                    "description": "Why the refund was requested.",
                    "type": "string"
                },
                "refundId": {
                    "description": "The refund request's ID.",
                    "type": "string"
                },
                "status": {
                    "description": "The refund status (`APPROVED`, `REJECTED`, or `COMPLETED`) — check this, not just the HTTP status, since a rejection is still a 201 response.",
                    "type": "string"
                }
            }
        },
        "http.RequestRefundRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "The amount to refund. Must be a positive integer not exceeding the original payment amount.",
                    "type": "integer"
                },
                "reason": {
                    "description": "Why the refund is being requested.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token issued by `POST /auth/sign-in`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}